SpriteAnimator: move to a separate category
It does not render anything, so it should not be in render/
